api user routes, auth routes added
