+++ server/api/users.js
@@ -1,61 +1,27 @@
 const router = require('express').Router()
 const {User} = require('../db/models')
+const {adminsOnly, adminOrUser} = require('../auth/privileges')
 module.exports = router
 
-const adminsOnly = (req, res, next) => {
-  if (req.user && !req.user.isAdmin) {
-    const err = new Error('Not An Admin')
-    err.status = 401
-    return next(err)
-  }
-  next()
-}
-
-const adminOrUser = (req, res, next) => {
-  if (
-    req.user &&
-    req.user.id !== req.params.id &&
-    (req.user && !req.user.isAdmin)
-  ) {
-    const err = new Error('Not An Admin')
-    err.status = 401
-    return next(err)
-  }
-  next()
-}
-
-router.put('/:id', adminOrUser, async (req, res, next) => {
+router.get('/', adminsOnly, async (req, res, next) => {
   try {
-    const {
-      email,
-      password,
-      name,
-      sex,
-      dietaryPreference,
-      birthdate,
-      height,
-      weight,
-      nutritionStatus
-    } = req.body
+    const users = await User.findAll({
+      attributes: ['id', 'email']
+    })
+    res.status(200).json(users)
+  } catch (err) {
+    next(err)
+  }
+})
+
+router.get('/:id', adminsOnly, async (req, res, next) => {
+  try {
     const user = await User.findByPk(req.params.id)
     if (user) {
-      await user.update({
-        email,
-        password,
-        name,
-        sex,
-        dietaryPreference,
-        birthdate,
-        height,
-        weight,
-        nutritionStatus
-      })
       res.status(200).json(user)
-    } else {
-      res.status(404).send('User not found.')
     }
-  } catch (error) {
-    next(error)
+  } catch (err) {
+    next(err)
   }
 })
 
@@ -73,58 +39,73 @@ router.delete('/:id', adminOrUser, async (req, res, next) => {
   }
 })
 
-router.post('/', async (req, res, next) => {
-  try {
-    const {
-      email,
-      password,
-      name,
-      sex,
-      dietaryPreference,
-      birthdate,
-      height,
-      weight,
-      nutritionStatus
-    } = req.body
-    const newUser = await User.create({
-      email,
-      password,
-      name,
-      sex,
-      dietaryPreference,
-      birthdate,
-      height,
-      weight,
-      nutritionStatus
-    })
-    if (newUser) {
-      res.status(201).json(newUser)
-    } else {
-      res.status(404).send('Could not create new user.')
-    }
-  } catch (error) {
-    next(error)
-  }
-})
+/* Updating profile */
+// router.put('/:id', adminOrUser, async (req, res, next) => {
+//   try {
+//     const {
+//       email,
+//       password,
+//       name,
+//       sex,
+//       dietaryPreference,
+//       birthdate,
+//       height,
+//       weight,
+//       nutritionStatus
+//     } = req.body
+//     const user = await User.findByPk(req.params.id)
+//     if (user) {
+//       await user.update({
+//         email,
+//         password,
+//         name,
+//         sex,
+//         dietaryPreference,
+//         birthdate,
+//         height,
+//         weight,
+//         nutritionStatus
+//       })
+//       res.status(200).json(user)
+//     } else {
+//       res.status(404).send('User not found.')
+//     }
+//   } catch (error) {
+//     next(error)
+//   }
+// })
 
-router.get('/:id', adminOrUser, async (req, res, next) => {
-  try {
-    const user = await User.findByPk(req.params.id)
-    if (user) {
-      res.status(200).json(user)
-    }
-  } catch (err) {
-    next(err)
-  }
-})
-
-router.get('/', adminsOnly, async (req, res, next) => {
-  try {
-    const users = await User.findAll({
-      attributes: ['id', 'email']
-    })
-    res.status(200).json(users)
-  } catch (err) {
-    next(err)
-  }
-})
+/* Making profile */
+// router.post('/', async (req, res, next) => {
+//   try {
+//     const {
+//       email,
+//       password,
+//       name,
+//       sex,
+//       dietaryPreference,
+//       birthdate,
+//       height,
+//       weight,
+//       nutritionStatus
+//     } = req.body
+//     const newUser = await User.create({
+//       email,
+//       password,
+//       name,
+//       sex,
+//       dietaryPreference,
+//       birthdate,
+//       height,
+//       weight,
+//       nutritionStatus
+//     })
+//     if (newUser) {
+//       res.status(201).json(newUser)
+//     } else {
+//       res.status(404).send('Could not create new user.')
+//     }
+//   } catch (error) {
+//     next(error)
+//   }
+// })
